Do not trust getpwuid_r to work on Emscripten
Autoconf test for getpwuid_r passes on Emscripten but the implementation
is broken, throwing TODO in the latest release. We don't need it when
running on browser so just disable it for now.

This will break looking dictionaries from user's home directory on
Emscripten builds. Hopefully no-one is doing such thing.