Remove autoconf tests for malloc and realloc (#49)
Remove tests for GNU compatible malloc and realloc functions.
Realloc is not used in the code at all. Malloc is used but only
with arguments that are always non-zero. Thus GNU compatibility
is not relevant anymore. These tests may sometimes break for
unrelated reasons, causing replacement malloc function to be
used which in turn causes compilation failures in C++ code